RF: Plain move. @build_doc is now base, not utils
