+++ datalad/distribution/clone.py
@@ -17,7 +17,7 @@ from os.path import exists
 
 from datalad.interface.base import Interface
 from datalad.interface.utils import eval_results
-from datalad.interface.utils import build_doc
+from datalad.interface.base import build_doc
 from datalad.interface.results import get_status_dict
 from datalad.interface.common_opts import location_description
 # from datalad.interface.common_opts import git_opts
