+++ datalad/distribution/create_test_dataset.py
@@ -26,7 +26,7 @@ from datalad.support.constraints import EnsureStr, EnsureNone, EnsureInt
 from datalad.support.gitrepo import GitRepo
 from datalad.support.annexrepo import AnnexRepo
 from datalad.interface.base import Interface
-from datalad.interface.utils import build_doc
+from datalad.interface.base import build_doc
 
 lgr = logging.getLogger('datalad.distribution.tests')
