+++ datalad/distribution/subdatasets.py
@@ -23,7 +23,7 @@ from git import GitConfigParser
 
 from datalad.interface.base import Interface
 from datalad.interface.utils import eval_results
-from datalad.interface.utils import build_doc
+from datalad.interface.base import build_doc
 from datalad.interface.results import get_status_dict
 from datalad.support.constraints import EnsureBool
 from datalad.support.constraints import EnsureStr
