+++ datalad/interface/common_opts.py
@@ -12,9 +12,11 @@
 
 __docformat__ = 'restructuredtext'
 
+from datalad.interface.results import known_result_xfms
 from datalad.support.param import Parameter
 from datalad.support.constraints import EnsureInt, EnsureNone, EnsureStr
 from datalad.support.constraints import EnsureChoice
+from datalad.support.constraints import EnsureCallable
 
 
 location_description = Parameter(
@@ -228,3 +230,59 @@ with_plugin_opt = Parameter(
     one plugin call. Plugins are called in the order defined by this list.
     PY][CMD: This option can be given more than once to run multiple plugins
     in the order in which they are given. CMD]""")
+
+# define parameters to be used by eval_results to tune behavior
+# Note: This is done outside eval_results in order to be available when building
+# docstrings for the decorated functions
+# TODO: May be we want to move them to be part of the classes _params. Depends
+# on when and how eval_results actually has to determine the class.
+# Alternatively build a callable class with these to even have a fake signature
+# that matches the parameters, so they can be evaluated and defined the exact
+# same way.
+
+eval_params = dict(
+    return_type=Parameter(
+        doc="""return value behavior switch. If 'item-or-list' a single
+        value is returned instead of a one-item return value list, or a
+        list in case of multiple return values. `None` is return in case
+        of an empty list.""",
+        constraints=EnsureChoice('generator', 'list', 'item-or-list')),
+    result_filter=Parameter(
+        doc="""if given, each to-be-returned
+        status dictionary is passed to this callable, and is only
+        returned if the callable's return value does not
+        evaluate to False or a ValueError exception is raised. If the given
+        callable supports `**kwargs` it will additionally be passed the
+        keyword arguments of the original API call.""",
+        constraints=EnsureCallable() | EnsureNone()),
+    result_xfm=Parameter(
+        doc="""if given, each to-be-returned result
+        status dictionary is passed to this callable, and its return value
+        becomes the result instead. This is different from
+        `result_filter`, as it can perform arbitrary transformation of the
+        result value. This is mostly useful for top-level command invocations
+        that need to provide the results in a particular format. Instead of
+        a callable, a label for a pre-crafted result transformation can be
+        given.""",
+        constraints=EnsureChoice(*list(known_result_xfms.keys())) | EnsureCallable() | EnsureNone()),
+    result_renderer=Parameter(
+        doc="""format of return value rendering on stdout""",
+        constraints=EnsureChoice('default', 'json', 'json_pp', 'tailored') | EnsureNone()),
+    on_failure=Parameter(
+        doc="""behavior to perform on failure: 'ignore' any failure is reported,
+        but does not cause an exception; 'continue' if any failure occurs an
+        exception will be raised at the end, but processing other actions will
+        continue for as long as possible; 'stop': processing will stop on first
+        failure and an exception is raised. A failure is any result with status
+        'impossible' or 'error'. Raised exception is an IncompleteResultsError
+        that carries the result dictionaries of the failures in its `failed`
+        attribute.""",
+        constraints=EnsureChoice('ignore', 'continue', 'stop')),
+)
+eval_defaults = dict(
+    return_type='list',
+    result_filter=None,
+    result_renderer=None,
+    result_xfm=None,
+    on_failure='continue',
+)
