+++ datalad/support/sshrun.py
@@ -21,7 +21,7 @@ import sys
 
 from datalad.support.param import Parameter
 from datalad.interface.base import Interface
-from datalad.interface.utils import build_doc
+from datalad.interface.base import build_doc
 
 from datalad import ssh_manager
